+++ tests/tests.py
@@ -391,6 +391,10 @@ def test_unconditional_cut_back():
 	run('-u -5', 'unconditional-back.fastq', 'small.fastq')
 
 
+def test_unconditional_cut_both():
+	run('-u -5 -u 5', 'unconditional-both.fastq', 'small.fastq')
+
+
 def test_no_zero_cap():
 	run("--no-zero-cap -c -e 0.122 -a CGCCTTGGCCGTACAGCAG", "solid-no-zerocap.fastq", "solid.fastq")